# backend/PHASE10_LS_NMC_API_CONTRACTS.yaml
openapi: 3.0.3
info:
  title: LS-NMC Governance API
  version: 1.0.0
paths:
  /admin/meta-controller/status:
    get:
      summary: Meta-controller runtime status
      responses:
        '200':
          description: Status payload
  /admin/meta-controller/enable:
    post:
      summary: Enable or disable meta-controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Toggle accepted
  /admin/meta-controller/model/promote:
    post:
      summary: Promote model version to prod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model_version]
              properties:
                model_version:
                  type: integer
      responses:
        '200':
          description: Promotion applied
  /admin/meta-controller/model/rollback:
    post:
      summary: Roll back to previous prod model version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_version]
              properties:
                target_version:
                  type: integer
      responses:
        '200':
          description: Rollback applied
  /admin/meta-controller/knob-log:
    get:
      summary: Retrieve applied/deterministic/neural knob values by run
      parameters:
        - in: query
          name: run_id
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Knob audit rows
  /admin/meta-controller/incidents:
    get:
      summary: Retrieve neural guardrail incidents
      responses:
        '200':
          description: Incident rows
components:
  schemas:
    MetaKnobs:
      type: object
      required: [alpha, beta, gamma, p_mult, u_mult]
      properties:
        alpha:
          type: number
          minimum: 0.2
          maximum: 0.8
        beta:
          type: number
          minimum: 0.2
          maximum: 0.8
        gamma:
          type: number
          minimum: 1.0
          maximum: 2.0
        p_mult:
          type: number
          minimum: 0.8
          maximum: 1.3
        u_mult:
          type: number
          minimum: 0.8
          maximum: 1.3
    KnobComputationResult:
      allOf:
        - $ref: '#/components/schemas/MetaKnobs'
        - type: object
          required: [fallback_used, guardrail_passed]
          properties:
            fallback_used:
              type: boolean
            guardrail_passed:
              type: boolean
            model_version:
              type: integer
              nullable: true